+++ CHANGELOG.md
@@ -16,11 +16,13 @@ All notable changes to this project will be documented in this file.
   { ... }
 ]
 ```
+- Added warning types: INFO, WARNING, ERROR
 
 #### Changed
 - RpcConnector now stores cookie into file
 - Set CONNECTTIMEOUT and TIMEOUT in RpcConnector
 - Use new object perunFacility in LDAP to search information about facility
+- Configuration for warning on DS is now in module_perun.php
 
 ## [v3.2.1]
 #### Fixed